// src/interfaces/prescription.ts
export interface PrescriptionInterface {
  Rx_id: string;
  Rx_Number: string;
  Order_Number: string;
  Patient_Name: string;
  Prescriber_Name: string;
  Status: string;
  tracking_code: string;
  Tracking: string;
  Medication: string;
  Issue_Date: string;
  doctor: string;
  add_date: string;
  Prescription_Number: string;
  patient_first_name: string;
  patient_last_name: string;
  Product_Name: string;
  firstName: string;
  lastName: string;
  dateOfBirth: string;
  issue_date: string;
  prescriber_first_name: string;
  prescriber_last_name: string;
  drug_name: string;
  order_status_id: string;
  status_cn: string;
  Order_id: string;
  script_no: string;
  price: number;
  quantity: number;
  refills: number;
  comments: string;
  Invoice_Nbr: string;
  selectedMedication: {
    sig_dosing_instruction: string;
    Rx_Number: string;
    Order_Number: string;
    Patient_Name: string;
    Prescriber_Name: string;
    Status: string;
    tracking_code: string;
    Tracking: string;
    Medication: string;
    Issue_Date: string;
    Prescription_Number: string;
    patient_first_name: string;
    patient_last_name: string;
    Product_Name: string;
    firstName: string;
    lastName: string;
    dateOfBirth: string;
    rx_id: string;
    issue_date: string;
    prescriber_first_name: string;
    prescriber_last_name: string;
    drug_name: string;
    order_status_id: string;
    Order_id: string;
    short_name: string;
  };
}